Refactor submodule in `reader`: `transformer` → `formatter`
